ChatGPT + OCR бот/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
easyocr==1.7.1
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
openai==1.13.3
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
pyclipper==1.3.0.post5
pydantic==2.6.3
pydantic_core==2.16.3
pyTelegramBotAPI==4.16.1
python-bidi==0.4.2
python-dotenv==1.0.1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
sympy==1.12
telebot==0.0.5
tifffile==2023.7.10
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
